deattached observations from attachments
