+++ config/routes.rb
@@ -121,9 +121,8 @@ Rails.application.routes.draw do
 
     resources :issue_tokens, only: %i[show] do
       get :show_by_token
-      resources :observations, only: [:update] do
-        resources :attachments, only: [:create]
-      end
+      resources :attachments, only: [:create]
+      resources :observations, only: [:update]
     end
   end
