+++ spec/requests/api/issue_tokens_spec.rb
@@ -69,8 +69,16 @@ describe IssueToken do
 
     observations.each do |observation|
       seed = observation.observable
+      api_update_issue_token(
+        "/issue_tokens/#{issue_token.token}/observations/#{observation.id}",
+        type: 'observations',
+        id: observation.id,
+        attributes: { reply: 'Some reply here' }
+      )
+      expect(api_response.data.attributes.state).to eq('answered')
+
       api_create_issue_token(
-        "/issue_tokens/#{issue_token.token}/observations/#{observation.id}/attachments",
+        "/issue_tokens/#{issue_token.token}/attachments",
         type: 'attachments',
         relationships: { attached_to_seed: { data: { id: seed.id, type: 'natural_docket_seeds' } } },
         attributes: {
@@ -79,14 +87,6 @@ describe IssueToken do
           document_content_type: mime_for(:jpg)
         }
       )
-
-      api_update_issue_token(
-        "/issue_tokens/#{issue_token.token}/observations/#{observation.id}",
-        type: 'observations',
-        id: observation.id,
-        attributes: { reply: 'Some reply here' }
-      )
-      expect(api_response.data.attributes.state).to eq('answered')
     end
 
     api_get "/issues/#{issue.id}"
